Add “NoReloadingStrategy” for SecorConfig
Currently there’s a lot of time spent reloading the configuration file
from disk, which is a pretty nifty mechanism, but it only applies to
a certain (not really well-defined) set of configuration parameters, and
it should either apply to all configuration parameters, or non. The
latter is covered by this commit.