Pluggable stats reporters
This commit introduces the concept of StatsReporter, which is a really
simple interface whose sole purpose is to push Stat objects to an
external service of choice.

This required extracting functionality from ProgressMonitor to separate
classes, and a ReporterFactory that given a proper configuration creates
StatReporter objects.

It utilises the fact that Apache Commons Configuration has support for
creating partial configuration objects using a key prefix. In this case
the prefix is `secor.stats`, and the `reporters` suffix specifies a list
of key-value pairs (separated by an equals sign (=)). The key is
considered an ID and the value should be a class implementing
StatsReporter. The ID is then used to extract configuration which will
be passed when creating an instance of said “value” class.

So for example in the OpenTSDB case the key is `opentsdb` and the
hostport configuration option is put under the `secor.stats.opentsdb`
namespace.